+++ src/syslogmp/stream.py
@@ -10,17 +10,18 @@ Treat binary data as a stream and provide methods to read from it.
 """
 
 from itertools import islice, takewhile
+from typing import Callable, Iterator
 
 
 class Stream:
-    def __init__(self, data):
+    def __init__(self, data: bytes) -> None:
         self.iterator = iter(data)
 
-    def read(self, n):
+    def read(self, n: int) -> bytes:
         """Return the next `n` bytes."""
         return bytes(islice(self.iterator, n))
 
-    def read_until(self, stop_byte):
+    def read_until(self, stop_byte: bytes) -> bytes:
         """Return bytes until the first occurrence of the stop byte.
 
         The stop byte is not returned, but silently dropped from the
@@ -29,12 +30,12 @@ class Stream:
         predicate = create_match_predicate(stop_byte)
         return bytes(takewhile(predicate, self.iterator))
 
-    def read_until_inclusive(self, stop_byte):
+    def read_until_inclusive(self, stop_byte: bytes) -> bytes:
         """Return bytes until, and including, the first occurrence of
         the stop byte.
         """
 
-        def inner():
+        def inner() -> Iterator[int]:
             predicate = create_match_predicate(stop_byte)
             for code_point in self.iterator:
                 yield code_point
@@ -43,11 +44,11 @@ class Stream:
 
         return bytes(inner())
 
-    def read_remainder(self):
+    def read_remainder(self) -> bytes:
         """Return all remaining bytes."""
         return bytes(self.iterator)
 
 
-def create_match_predicate(value_to_match):
-    value_to_match = ord(value_to_match)
-    return lambda cp: cp != value_to_match
+def create_match_predicate(value_to_match: bytes) -> Callable[[int], bool]:
+    code_point = ord(value_to_match)
+    return lambda cp: cp != code_point